push draft

+++ mobile-be/src/main/java/mobile_be/mobile_be/Repository/NoticeRepository.java
@@ -3,5 +3,5 @@ package mobile_be.mobile_be.Repository;
 import mobile_be.mobile_be.Model.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoticeRepository extends JpaRepository<Notice, Long> {
+public interface NoticeRepository extends JpaRepository<Notice, Integer> {
 }
